refactor : Rename 파일명 변경
talearnt 명칭을 taleant로 변경
